Refactoring cosmos test chain impls
