+++ crates/cosmos/cosmos-test-components/src/chain/impls/chain_id.rs
@@ -1,5 +1,5 @@
 use hermes_relayer_components::chain::traits::types::chain_id::HasChainIdType;
-use hermes_test_components::chain_driver::traits::build::chain_id::ChainIdFromStringBuilder;
+use hermes_test_components::chain::traits::chain_id::ChainIdFromStringBuilder;
 use hermes_test_components::chain_driver::traits::types::chain::HasChainType;
 use ibc_relayer_types::core::ics24_host::identifier::ChainId;
