+++ crates/cosmos/cosmos-test-components/src/chain/impls/queries/mod.rs
@@ -1,0 +1,1 @@
+pub mod balance;
