+++ crates/cosmos/cosmos-test-components/src/chain/impls/types/denom.rs
@@ -1,7 +1,7 @@
 use cgp_core::Async;
-use hermes_test_components::chain_driver::traits::types::denom::ProvideDenomType;
+use hermes_test_components::chain::traits::types::denom::ProvideDenomType;
 
-use crate::chain_driver::types::denom::Denom;
+use crate::chain::types::denom::Denom;
 
 pub struct ProvideIbcDenom;
