+++ crates/cosmos/cosmos-test-components/src/chain/impls/types/mod.rs
@@ -1,0 +1,4 @@
+pub mod address;
+pub mod amount;
+pub mod denom;
+pub mod wallet;
