+++ crates/cosmos/cosmos-test-components/src/chain/impls/types/wallet.rs
@@ -1,12 +1,12 @@
 use hermes_relayer_components::transaction::traits::types::HasSignerType;
-use hermes_test_components::chain_driver::traits::types::address::HasAddressType;
-use hermes_test_components::chain_driver::traits::types::tx_context::HasTxContextType;
-use hermes_test_components::chain_driver::traits::types::wallet::{
+use hermes_test_components::chain::traits::types::address::HasAddressType;
+use hermes_test_components::chain::traits::types::tx_context::HasTxContextType;
+use hermes_test_components::chain::traits::types::wallet::{
     HasWalletType, ProvideWalletType, WalletSignerProvider,
 };
 use ibc_relayer::keyring::Secp256k1KeyPair;
 
-use crate::chain_driver::types::wallet::CosmosTestWallet;
+use crate::chain::types::wallet::CosmosTestWallet;
 
 pub struct ProvideCosmosTestWallet;
