+++ crates/cosmos/cosmos-test-components/src/chain/mod.rs
@@ -1,0 +1,2 @@
+pub mod impls;
+pub mod types;
