+++ crates/cosmos/cosmos-test-components/src/chain_driver/impls/mod.rs
@@ -1,13 +1,5 @@
-pub mod address;
 pub mod amount;
-pub mod chain_id;
-pub mod convert_ibc_amout;
-pub mod denom;
 pub mod deposit_proposal;
-pub mod ibc_transfer_timeout;
-pub mod messages;
 pub mod proposal_status;
-pub mod query_balance;
 pub mod store_wasm_client;
 pub mod vote_proposal;
-pub mod wallet;
