+++ crates/cosmos/cosmos-test-components/src/chain_driver/mod.rs
@@ -1,3 +1,2 @@
 pub mod impls;
 pub mod traits;
-pub mod types;
