+++ crates/cosmos/cosmos-test-components/src/lib.rs
@@ -3,4 +3,5 @@
 extern crate alloc;
 
 pub mod bootstrap;
+pub mod chain;
 pub mod chain_driver;
