+++ crates/test/test-components/src/chain/traits/chain_id.rs
@@ -5,9 +5,6 @@ use hermes_relayer_components::chain::types::aliases::ChainIdOf;
 use crate::chain_driver::traits::types::chain::HasChainType;
 
 #[derive_component(ChainIdFromStringBuilderComponent, ChainIdFromStringBuilder<Chain>)]
-pub trait CanBuildChainIdFromString: HasChainType
-where
-    Self::Chain: HasChainIdType,
-{
-    fn build_chain_id_from_string(chain_id: &str) -> ChainIdOf<Self::Chain>;
+pub trait CanBuildChainIdFromString: HasChainIdType {
+    fn build_chain_id_from_string(chain_id: &str) -> Self::ChainId;
 }
