+++ crates/test/test-components/src/chain_driver/traits/mod.rs
@@ -1,4 +1,3 @@
 pub mod assert;
-pub mod build;
 pub mod fields;
 pub mod types;
